More refactoring of MAGE-TAB parsing functions and submission type variable
